Iteration Trigger Update
The looping of the animation once point B is reached has successfully been implemented. All that is required now is to install a timer for each cycle and ensure this resets through each iter.

Also it is fundamental to work out what the variable looks like for the Apache StopWatch feature.